requirements.txt
astroid==2.11.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
dill==0.3.4
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
pep8==1.7.1
platformdirs==2.5.2
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.13.5
pyparsing==3.0.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
